Issue #24858 move annotations for fulley annotative queries
