+++ dev/io.openliberty.data.internal.persistence/src/io/openliberty/data/internal/persistence/Condition.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2022 IBM Corporation and others.
+ * Copyright (c) 2022,2023 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License 2.0
  * which accompanies this distribution, and is available at
@@ -14,8 +14,8 @@ package io.openliberty.data.internal.persistence;
 
 import com.ibm.websphere.ras.annotation.Trivial;
 
+import io.openliberty.data.repository.Compare;
 import jakarta.data.exceptions.MappingException;
-import jakarta.data.repository.Compare;
 
 /**
  */
