+++ dev/io.openliberty.data.internal_fat/test-applications/DataTestApp/src/test/jakarta/data/web/Shipments.java
@@ -15,14 +15,15 @@ package test.jakarta.data.web;
 import java.time.OffsetDateTime;
 import java.util.stream.Stream;
 
-import jakarta.data.repository.Compare;
-import jakarta.data.repository.Delete;
-import jakarta.data.repository.Filter;
 import jakarta.data.repository.OrderBy;
 import jakarta.data.repository.Param;
 import jakarta.data.repository.Repository;
-import jakarta.data.repository.Select;
-import jakarta.data.repository.Update;
+
+import io.openliberty.data.repository.Compare;
+import io.openliberty.data.repository.Delete;
+import io.openliberty.data.repository.Filter;
+import io.openliberty.data.repository.Select;
+import io.openliberty.data.repository.Update;
 
 /**
  *
