+++ dev/io.openliberty.data/.gitignore
@@ -1,0 +1,2 @@
+/bin/
+/generated/
