+++ dev/io.openliberty.data/src/io/openliberty/data/repository/Filter.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2022,2023 IBM Corporation and others.
+ * Copyright (c) 2023 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License 2.0
  * which accompanies this distribution, and is available at
@@ -10,7 +10,7 @@
  * Contributors:
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
-package jakarta.data.repository;
+package io.openliberty.data.repository;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Repeatable;
@@ -18,32 +18,36 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-/**
- * TODO propose in Jakarta Data.
- * The {@link Filter &#64;Filter} annotation can be added to provide conditions.
- */
-@Repeatable(Update.List.class)
+@Repeatable(Filter.List.class)
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.METHOD)
-public @interface Update {
+public @interface Filter {
+    Type as() default Type.AND;
 
-    String attr();
+    String by();
 
-    Operation op() default Operation.Assign;
+    boolean ignoreCase() default false;
 
-    String param() default "";
+    Compare op() default Compare.Equal;
+
+    String[] param() default {};
 
     // Single quotes are automatically added to the beginning and end of values
     // unless the value starts with a number or single quote
     String[] value() default {};
 
     /**
-     * Enables multiple {@link Update}
+     * Enables multiple {@link Filter}
      * annotations on the same type.
      */
     @Retention(RetentionPolicy.RUNTIME)
     @Target(ElementType.METHOD)
     public @interface List {
-        Update[] value();
+        Filter[] value();
+    }
+
+    public enum Type {
+        AND,
+        OR
     }
 }
