+++ dev/io.openliberty.data/src/io/openliberty/data/repository/Update.java
@@ -10,19 +10,39 @@
  * Contributors:
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
-package jakarta.data.repository;
+package io.openliberty.data.repository;
 
 import java.lang.annotation.ElementType;
+import java.lang.annotation.Repeatable;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 /**
- * TODO propose this in Jakarta Data.
- * Annotates a repository method to designate it as an exists operation.
  * The {@link Filter &#64;Filter} annotation can be added to provide conditions.
  */
+@Repeatable(Update.List.class)
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.METHOD)
-public @interface Exists {
+public @interface Update {
+
+    String attr();
+
+    Operation op() default Operation.Assign;
+
+    String param() default "";
+
+    // Single quotes are automatically added to the beginning and end of values
+    // unless the value starts with a number or single quote
+    String[] value() default {};
+
+    /**
+     * Enables multiple {@link Update}
+     * annotations on the same type.
+     */
+    @Retention(RetentionPolicy.RUNTIME)
+    @Target(ElementType.METHOD)
+    public @interface List {
+        Update[] value();
+    }
 }
