--- dev/io.openliberty.jakarta.data.1.0/src/jakarta/data/repository/Filter.java
@@ -1,53 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2023 IBM Corporation and others.
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License 2.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-2.0/
- *
- * SPDX-License-Identifier: EPL-2.0
- *
- * Contributors:
- *     IBM Corporation - initial API and implementation
- *******************************************************************************/
-package jakarta.data.repository;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Repeatable;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Repeatable(Filter.List.class)
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.METHOD)
-public @interface Filter {
-    Type as() default Type.AND;
-
-    String by();
-
-    boolean ignoreCase() default false;
-
-    Compare op() default Compare.Equal;
-
-    String[] param() default {};
-
-    // Single quotes are automatically added to the beginning and end of values
-    // unless the value starts with a number or single quote
-    String[] value() default {};
-
-    /**
-     * Enables multiple {@link Filter}
-     * annotations on the same type.
-     */
-    @Retention(RetentionPolicy.RUNTIME)
-    @Target(ElementType.METHOD)
-    public @interface List {
-        Filter[] value();
-    }
-
-    public enum Type {
-        AND,
-        OR
-    }
-}
